+++ CHANGELOG.md
@@ -31,6 +31,10 @@
 - Removed now non-existent app handle from settings, and `getAppHandle()` from handles service. Unlike OAuth Client, the underlying auth plugin doesn’t have a concept of “apps.”
 - Removed display of orders by a payee on a payee’s user page
 - Removed support for configuring fees visually via the settings in the control panel
+- Removed `marketplace/accounts/create-logout-link` action, which has no role in new account flow
+
+## Fixed
+- Fixed internal name of 
 
 ***